Freq by compound head word form, semtype by lemma
noun_semtype.data can be used to split compounded words. This data is indexed by lemma. Words are lemmatized by Frog so this can be used to search for a split.

Frequencies however are determined by (inflected) word form. By "substracting" the lemma compound parts from the complete word, the word form of the compound head can be determined.